refactor(anova): group DoE/ANOVA modules into agent_eval.anova package
matrix.py, composite.py, archive.py, anova_runner.py and stats/ were top-level
modules in agent_eval/ alongside the core harness. They are all one concern
(eval-anova / Design-of-Experiments), so move them into a dedicated
agent_eval/anova/ package and rename anova_runner.py → runner.py:

    agent_eval/anova/{matrix,composite,archive,runner}.py
    agent_eval/anova/stats/{anova,pareto}.py

Every importer is inside this PR (eval/harbor-maas-v1/driver.py, the
eval-anova skill scripts, and the anova tests) — no core-harness or external
consumer imports these, so the move is self-contained with no back-compat
shims. Doing it now (before these modules exist on main) avoids a future
breaking relocation. Updated all imports, the runner's repo-root depth, and the
CLAUDE.md architecture tree.
