+++ agent_eval/anova/__init__.py
@@ -1,0 +1,6 @@
+"""eval-anova: Design-of-Experiments matrix testing + ANOVA statistics.
+
+Matrix design (matrix), composite scoring (composite), the legacy per-cell
+executor (runner), git-backed archival (archive), and the statistics package
+(stats: repeated-measures / mixed-effects / one-way ANOVA + Pareto frontier).
+"""
