+++ agent_eval/anova/stats/__init__.py
@@ -3,10 +3,10 @@
 Optional dependency: install with `pip install agent-eval-harness[anova]`.
 """
 
-from agent_eval.stats.pareto import pareto_frontier
+from agent_eval.anova.stats.pareto import pareto_frontier
 
 try:
-    from agent_eval.stats.anova import (
+    from agent_eval.anova.stats.anova import (
         mixed_effects_anova,
         one_way_anova,
         repeated_measures_anova,
